+++ src/socket_server.h
@@ -14,9 +14,9 @@
 struct socket_server;
 
 struct socket_message {
-	int id;
+    size_t id;
 	uintptr_t opaque;
-	int ud;	// for accept, ud is listen id ; for data, ud is size of data 
+    size_t ud;	// for accept, ud is listen id ; for data, ud is size of data 
 	char * data;
 };
